Moved TasksActionBar into HeaderBar.
