+++ src/js/pages/Tasks.jsx
@@ -2,19 +2,15 @@ import { withStyles } from '@mui/styles';
 import React, { useEffect, useState } from 'react';
 import { Helmet } from 'react-helmet-async';
 import styled from 'styled-components';
-import SearchBar2024 from '../common/components/Search/SearchBar2024';
 import { renderLog } from '../common/utils/logging';
 import PersonSummaryHeader from '../components/Person/PersonSummaryHeader';
 import PersonSummaryRow from '../components/Person/PersonSummaryRow';
-import { ActionBarItem, ActionBarSection, SearchBarWrapper } from '../components/Style/actionBarStyles';
-import { SpanWithLinkStyle } from '../components/Style/linkStyles';
 import { PageContentContainer } from '../components/Style/pageLayoutStyles';
 import TaskListForPerson from '../components/Task/TaskListForPerson';
 import webAppConfig from '../config';
 import { useConnectAppContext, useConnectDispatch } from '../contexts/ConnectAppContext';
 import { isSearchTextFoundInPerson } from '../controllers/PersonController';
 import { isSearchTextFoundInTask } from '../controllers/TaskController';
-import { viewerCanSeeOrDo } from '../models/AuthModel';
 import capturePersonListRetrieveData from '../models/capturePersonListRetrieveData';
 import {
   captureTaskDefinitionListRetrieveData, captureTaskGroupListRetrieveData, captureTaskStatusListRetrieveData,
@@ -26,8 +22,8 @@ import { alphabetizePeoplesObject } from '../utils/utilities';
 
 const Tasks = () => {
   renderLog('Tasks');  // Set LOG_RENDER_EVENTS to log all renders
-  const { apiDataCache, setAppContextValue } = useConnectAppContext();
-  const { allPeopleCache, allTaskDefinitionsCache, allTasksCache, viewerAccessRights } = apiDataCache;
+  const { apiDataCache, getAppContextValue } = useConnectAppContext();
+  const { allPeopleCache, allTaskDefinitionsCache, allTasksCache } = apiDataCache;
   const dispatch = useConnectDispatch();
 
   const [personIdsList, setPersonIdsList] = useState([]);
@@ -36,15 +32,6 @@ const Tasks = () => {
   const [showCompletedTasks, setShowCompletedTasks] = useState(false);
   const [taskListByPersonId, setTaskListByPersonId] = useState({});
   const [taskDefinitionList, setTaskDefinitionList] = useState([]);
-
-  const clearFunction = () => {
-    setSearchText('');
-  };
-
-  const searchFunction = (incomingSearchText) => {
-    // console.log('AddTeamDrawerMainContent searchFunction incomingSearchText: ', incomingSearchText);
-    setSearchText(incomingSearchText);
-  };
 
   const personListRetrieveResults = useFetchData(['person-list-retrieve'], {}, METHOD.GET);
   useEffect(() => {
@@ -109,11 +96,6 @@ const Tasks = () => {
     // console.log('=== taskListByPersonIdTemp:', taskListByPersonIdTemp);
   }, [allPeopleCache, allTasksCache]);
 
-  const addTeamMemberClick = () => {
-    setAppContextValue('addPersonDrawerOpen', true);
-    setAppContextValue('AddPersonDrawerLabel', 'Add Person');
-  };
-
   const showPerson = (person) => {
     if (!person || !person.personId < 0) return false; // Invalid person or personId
     if (searchText) {
@@ -155,6 +137,15 @@ const Tasks = () => {
     }
   };
 
+  useEffect(() => {
+    if (getAppContextValue('tasksActionBarSearchText') !== searchText) {
+      setSearchText(getAppContextValue('tasksActionBarSearchText'));
+    }
+    if (getAppContextValue('tasksActionBarShowCompletedTasks') !== showCompletedTasks) {
+      setShowCompletedTasks(getAppContextValue('tasksActionBarShowCompletedTasks'));
+    }
+  }, [getAppContextValue]);
+
   const teamId = 0;  // hack 1/15/25
   return (
     <div>
@@ -169,39 +160,7 @@ const Tasks = () => {
         {/* browser.js:38 Uncaught Invariant Violation: Only elements types base, body, head, html, link, meta, noscript, script, style, title, Symbol(react.fragment) are allowed. Helmet does not support rendering <[object Object]> elements. Refer to our API for more information. */}
       </Helmet>
       <PageContentContainer>
-        <h1>Tasks</h1>
-        <TasksActionBarWrapper>
-          <SearchBarWrapper>
-            <SearchBar2024
-              clearFunction={clearFunction}
-              placeholder="Search existing tasks"
-              searchFunction={searchFunction}
-              searchUpdateDelayTime={0}
-            />
-          </SearchBarWrapper>
-          <ActionBarSection>
-            <ActionBarItem>
-              {showCompletedTasks ? (
-                <SpanWithLinkStyle onClick={() => setShowCompletedTasks(false)}>
-                  Hide completed tasks
-                </SpanWithLinkStyle>
-              ) : (
-                <SpanWithLinkStyle onClick={() => setShowCompletedTasks(true)}>
-                  Show completed tasks
-                </SpanWithLinkStyle>
-              )}
-            </ActionBarItem>
-          </ActionBarSection>
-          <ActionBarSection>
-            {viewerCanSeeOrDo(['canAddTeamMemberAnyTeam'], viewerAccessRights) && (
-              <ActionBarItem>
-                <SpanWithLinkStyle onClick={() => addTeamMemberClick()}>
-                  Add team member
-                </SpanWithLinkStyle>
-              </ActionBarItem>
-            )}
-          </ActionBarSection>
-        </TasksActionBarWrapper>
+        <ActionBarWrapperSpacer />
         <PersonSummaryHeader />
         {taskListByPersonId && selectedPersonList.map((person) => {
           const showPersonResults = showPerson(person);
@@ -239,14 +198,11 @@ const styles = (theme) => ({
   },
 });
 
-const OnePersonWrapper = styled('div')`
+const ActionBarWrapperSpacer = styled('div')`
+  margin-top: 60px;
 `;
 
-const TasksActionBarWrapper = styled('div')`
-  align-items: center;
-  display: flex;
-  justify-content: flex-start;
-  margin-top: 40px;  // Temporary hack
+const OnePersonWrapper = styled('div')`
 `;
 
 export default withStyles(styles)(Tasks);
